# microservices/lectures/lectures/08_auto_generating_HTTP_clients_from_Swagger_files/models/validation_error.go
// Code generated by go-swagger; DO NOT EDIT.

package models

// This file was generated by the swagger tool.
// Editing this file might prove futile when you re-run the swagger generate command

import (
	"context"

	"github.com/go-openapi/strfmt"
	"github.com/go-openapi/swag"
)

// ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError ValidationError is a collection of validation error messages.
//
// swagger:model ValidationError
type ValidationError struct {

	// messages
	Messages []string `json:"messages"`
}

// Validate validates this validation error
func (m *ValidationError) Validate(formats strfmt.Registry) error {
	return nil
}

// ContextValidate validates this validation error based on context it is used
func (m *ValidationError) ContextValidate(ctx context.Context, formats strfmt.Registry) error {
	return nil
}

// MarshalBinary interface implementation
func (m *ValidationError) MarshalBinary() ([]byte, error) {
	if m == nil {
		return nil, nil
	}
	return swag.WriteJSON(m)
}

// UnmarshalBinary interface implementation
func (m *ValidationError) UnmarshalBinary(b []byte) error {
	var res ValidationError
	if err := swag.ReadJSON(b, &res); err != nil {
		return err
	}
	*m = res
	return nil
}